requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0
numpy==2.1.2
pandas==2.2.3
scikit-learn==1.5.2
hdbscan==0.8.38
datasketch==1.6.5
pydantic==2.9.2
prometheus-client==0.21.0
sentence-transformers==3.0.1
